Add support for drop-in patches
See src-apply/README.md for full details. The goal is to be able to
apply our own patches to the musl-cross-make source tree as necessary,
without a risk of failure when those patches fail to apply after future
upstream commits. To do this, create a 'try-apply.sh' script which
tests patch --dry-run first, and then warns if it will fail, before
fully applying the patch.

The primary forseeable usage of these patches is support for newer
versions of toolchain packages not yet supported by upstream. Again,
the README.md file in the 'src-apply' directory contains more details
about this.